requirements.txt
pymongo==4.7.3
openai==1.35.0
azure-ai-vision-imageanalysis==1.0.0
azure-identity==1.17.1
python-dotenv==1.0.1
pypdf==4.3.1
azure-functions==1.24.0
